eval code

+++ eval.py
@@ -17,9 +17,9 @@ config.test_dir = os.path.join(config.test_dir, config.phone_model, "patches/*")
 files = sorted(glob(config.test_dir))
 num_samples = int(len(files) / 3)
 
-gts = np.zeros((1, 100, 100, 3))
-inputs = np.zeros((1, 100, 100, 3))
-outputs = np.zeros((1, 100, 100, 3))
+gts = np.zeros((1, 100, 100, 3), dtype=np.float32)
+inputs = np.zeros((1, 100, 100, 3), dtype=np.float32)
+outputs = np.zeros((1, 100, 100, 3), dtype=np.float32)
 for i in range(1):
     print(files[i * 3], files[i * 3 + 1], files[i * 3 + 2])
     gts[i] = imread(files[i * 3], mode="RGB")
